Move and merge turn processing functions
